services/media-service/pyproject.toml
[project]
name = "media-service"
version = "0.1.0"
description = "SMWCS Media Upload Service"
authors = [
    {name = "dee-mee",email = "derekmuriuki2@gmail.com"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "aioboto3 (>=15.5.0,<16.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "structlog (>=25.5.0,<26.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.8,<0.16.0)"
]
